+++ src/utils/neynar/getDataFromNeynar.test.ts
@@ -1,7 +1,7 @@
 import { version } from '../../version';
 import { NEYNAR_DEFAULT_API_KEY } from './frame/neynarFrameFunctions';
 import { getDataFromNeynar } from './getDataFromNeynar';
-  
+
 describe('getDataFromNeynar', () => {
   let fetchMock = jest.fn();
   let status = 200;
@@ -31,4 +31,3 @@ describe('getDataFromNeynar', () => {
     });
   });
 });
-
