+++ src/utils/neynar/user/types.ts
@@ -1,0 +1,17 @@
+export interface NeynarUserModel {
+  fid: number;
+  custody_address: string;
+  username: string;
+  display_name: string;
+  pfp_url: string;
+  profile: {
+    bio: {
+      text: string;
+    };
+  };
+  follower_count: number;
+  verifications: string[];
+}
+export interface NeynarBulkUserLookupModel {
+  users: NeynarUserModel[];
+}
